requirements.txt
langchain
llama-index
pymongo
nltk
Pillow
python-dotenv
